Moved `Image` to its own file.
